Fix subdir orderring (#1020)
