+++ flash/core/data/utilities/classification.py
@@ -11,13 +11,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-import re
 from enum import auto, Enum
 from functools import reduce
-from typing import Any, cast, Iterable, List, Optional, Tuple, Union
+from typing import Any, cast, List, Optional, Tuple, Union
 
 import numpy as np
 import torch
+
+from flash.core.data.utilities.sort import sorted_alphanumeric
 
 
 def _is_list_like(x: Any) -> bool:
@@ -33,24 +34,6 @@ def _as_list(x: Union[List, torch.Tensor, np.ndarray]) -> List:
     if torch.is_tensor(x) or isinstance(x, np.ndarray):
         return cast(List, x.tolist())
     return x
-
-
-def _convert(text: str) -> Union[int, str]:
-    return int(text) if text.isdigit() else text
-
-
-def _alphanumeric_key(key: str) -> List[Union[int, str]]:
-    return [_convert(c) for c in re.split("([0-9]+)", key)]
-
-
-def _sorted_nicely(iterable: Iterable[str]) -> Iterable[str]:
-    """Sort the given iterable in the way that humans expect. For example, given ``{"class_1", "class_11",
-    "class_2"}`` this returns ``["class_1", "class_2", "class_11"]``.
-
-    Copied from:
-    https://blog.codinghorror.com/sorting-for-humans-natural-sort-order/
-    """
-    return sorted(iterable, key=_alphanumeric_key)
 
 
 class TargetMode(Enum):
@@ -301,6 +284,6 @@ def get_target_details(targets: List[Any], target_mode: TargetMode) -> Tuple[Opt
             tokens = targets
 
         tokens = [token.strip() for token in tokens]
-        labels = list(_sorted_nicely(set(tokens)))
+        labels = list(sorted_alphanumeric(set(tokens)))
         num_classes = len(labels)
     return labels, num_classes
